move GetKubeVersion to controllerutil
